move country name generation to country_info module
