+++ nominatim/clicmd/setup.py
@@ -113,8 +113,8 @@ class SetupAll:
             database_import.create_search_indices(conn, args.config,
                                                   drop=args.no_updates)
             LOG.warning('Create search index for default country names.')
-            database_import.create_country_names(conn, tokenizer,
-                                                 args.config.LANGUAGES)
+            country_info.create_country_names(conn, tokenizer,
+                                              args.config.LANGUAGES)
             conn.commit()
             if args.no_updates:
                 freeze.drop_update_tables(conn)
